Unblock S.IO.Ports and S.IO.Pipes on Aot.
More stray P/Invokes...
