+++ src/System.IO.Pipes/tests/NamedPipeTests/NamedPipeTest.RunAsClient.Unix.cs
@@ -13,25 +13,6 @@ namespace System.IO.Pipes.Tests
 {
     public class NamedPipeTest_RunAsClient : RemoteExecutorTestBase
     {
-        [Fact]
-        [PlatformSpecific(TestPlatforms.Windows)]  // Uses P/Invokes
-        public async Task RunAsClient_Windows()
-        {
-            string pipeName = Path.GetRandomFileName();
-            using (var server = new NamedPipeServerStream(pipeName))
-            using (var client = new NamedPipeClientStream(".", pipeName, PipeDirection.InOut, PipeOptions.None, TokenImpersonationLevel.Impersonation))
-            {
-                Task serverTask = server.WaitForConnectionAsync();
-
-                client.Connect();
-                await serverTask;
-
-                bool ran = false;
-                server.RunAsClient(() => ran = true);
-                Assert.True(ran, "Expected delegate to have been invoked");
-            }
-        }
-
         [DllImport("libc", SetLastError = true)]
         internal static extern unsafe int seteuid(uint euid);
